Updated user login and user creation
Made a function in RestService for getting user data to load a user object (denoted by the user.cs file). Made get_user_data.php to get user data and load it. Also updated login page. Also think this is the first time the titlepage has made it to the main branch. The main app pages also has the order to make pages.