Add new user
Kyle assisted with user.php function to addUser, getting the query to function as intended. The rest of this commit contains the xamarin gui classes that handle pages for admins to create a new user account. It also contains the php files to handle what happens server side for when this is executed.